validate user on delete request

+++ part4/blog-list-app/controllers/blogs.js
@@ -37,10 +37,26 @@ blogsRouter.post('/', async (request, response) => {
     }
 })
 
+
 blogsRouter.delete('/:id', async (request, response) => {
-    await Blog.findByIdAndRemove(request.params.id)
-    response.status(204).end()
+
+    const blog = await Blog.findById(request.params.id)
+
+    const decodedToken = jwt.verify(request.token, process.env.SECRET)
+    if (!request.token || !decodedToken.id) {
+        return reponse.status(401).json({ error: 'token missing or invalid' })
+    }
+
+    const user = await User.findById(decodedToken.id)
+
+    if ( blog.user.toString() === user.id.toString() ) {
+        await Blog.findByIdAndRemove(request.params.id)
+        response.status(204).end()
+    } else {
+        response.status(401).json({ error: 'unauthorized'}).end()
+    }
 })
+
 
 blogsRouter.put('/:id', async (request, response, next) => {
     const body = request.body
